refactor(acp): route through invocation kernel
- Extract pure helpers (parseUnicliInvocation, runCommand,
  suggestCommands, summarizeResults, tokenize, coerce) to
  ./acp-helpers.ts so acp.ts stays under the complexity gate.
- runCommand now routes through buildInvocation/execute so ACP
  inherits ajv validation + hardening for free. Throws errors with a
  `.suggestion` property so callers can surface directional feedback
  on the ACP wire.
- Re-export the helpers from acp.ts for downstream callers.
- Only inject default `limit: 20` when the adapter declares it —
  ajv strict mode rejects undeclared properties.
